Update Push, Impact, Damage logic. Clean code

+++ HexEngine/Assets/Scripts/Entities/Models/Properties/Combat/DamagePropertySet.cs
@@ -13,7 +13,7 @@ public class DamagePropertySet
         IsTrueDamage = isTrueDamage;
     }
 
-    public int GetDamageValue(int impactValue)
+    public int GetDamageValue(float impactValue)
     {
         return Value + (int)Math.Ceiling(impactValue * ImpactValueModifier);
     }
